backend-app-api: split BackendInitializer.ts into focused internal modules
Extract three focused modules from BackendInitializer.ts to improve
readability and maintainability:

- instanceRegistry.ts: Process-level singleton managing backend instance
  shutdown hooks
- instanceMetadataServiceFactory.ts: Factory for the root instance
  metadata service
- featureDiscovery.ts: Type guards for backend feature classification
  (service factories, registrations, feature loaders)

Made-with: Cursor
Made-with: Cursor
